requirements.txt
flask==0.10.1
sqlalchemy==0.9.7
flask-sqlalchemy==2.0
sqlalchemy-migrate==0.9.2
requests==2.4.1
redis==2.10.3
python-gcm==0.2
